ai_agent/requirements.txt
langchain==0.1.0
langchain-openai==0.0.5
langchain-community==0.0.20
langgraph==0.0.20
openai==1.10.0
faiss-cpu==1.7.4
python-dotenv==1.0.0
beautifulsoup4==4.12.3
requests==2.31.0
lxml==5.1.0
pandas==2.1.4
numpy==1.26.3
tiktoken==0.5.2